requirements.txt
requests
python-dotenv
pandas
pyarrow
tqdm